+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 # Release Notes
 
 ### 0.6.2  (WIP)
+- #120 Adds command to Repeat Previous Search
 - #114 Adds "How To Use Koffee" link in settings window
 - #70 Changes undo move/copy to undo individual file operations for more correct behavior when folders were merged
 - #123 Fixes path suggestions not being sorted by MRU